[BE] Upload addAttachmentsToClasswork (#362)
* CU-8tx7vp - [BE] Update addAttachmentsToClassworkAssignment [BE] Update addAttachmentsToClassworkMaterial

* fix

* fix

* fix

* fix
