+++ packages/server/src/modules/classwork/classwork.service.ts
@@ -11,6 +11,7 @@ import {
 import { Course } from 'modules/academic/models/Course'
 import { AccountService } from 'modules/account/account.service'
 import { AuthService } from 'modules/auth/auth.service'
+import { FileStorageService } from 'modules/fileStorage/fileStorage.service'
 import { OrgService } from 'modules/org/org.service'
 // eslint-disable-next-line import/order
 import { ANY, Nullable, PageOptionsInput } from 'types'
@@ -18,6 +19,7 @@ import {
   UpdateClassworkMaterialInput,
   CreateClassworkAssignmentInput,
   CreateClassworkMaterialInput,
+  AddAttachmentsToClassworkInput,
 } from './classwork.type'
 import { Classwork, ClassworkType } from './models/Classwork'
 import { ClassworkAssignment } from './models/ClassworkAssignment'
@@ -40,6 +42,9 @@ export class ClassworkService {
 
     @InjectModel(Course)
     private readonly courseModel: ReturnModelType<typeof Course>,
+
+    @Inject(forwardRef(() => FileStorageService))
+    private readonly fileStorageService: FileStorageService,
 
     @Inject(forwardRef(() => AuthService))
     private readonly authService: AuthService,
@@ -416,8 +421,14 @@ export class ClassworkService {
   async addAttachmentsToClassworkMaterial(
     orgId: string,
     classworkMaterialId: string,
-    attachments?: string[],
+    attachmentsInput: AddAttachmentsToClassworkInput,
+    uploadedByAccountId: string,
   ): Promise<Nullable<DocumentType<ClassworkMaterial>>> {
+    const attachments = await this.uploadFilesAttachments(
+      orgId,
+      attachmentsInput,
+      uploadedByAccountId,
+    )
     return this.addAttachmentsToClasswork(
       orgId,
       classworkMaterialId,
@@ -679,8 +690,14 @@ export class ClassworkService {
   async addAttachmentsToClassworkAssignment(
     orgId: string,
     classworkAssignmentId: string,
-    attachments?: string[],
+    attachmentsInput: AddAttachmentsToClassworkInput,
+    uploadedByAccountId: string,
   ): Promise<Nullable<DocumentType<ClassworkAssignment>>> {
+    const attachments = await this.uploadFilesAttachments(
+      orgId,
+      attachmentsInput,
+      uploadedByAccountId,
+    )
     return this.addAttachmentsToClasswork(
       orgId,
       classworkAssignmentId,
@@ -700,6 +717,46 @@ export class ClassworkService {
       ClassworkType.Assignment,
       attachments,
     ) as Promise<Nullable<DocumentType<ClassworkAssignment>>>
+  }
+
+  async uploadFilesAttachments(
+    orgId: string,
+    attachmentsInput: AddAttachmentsToClassworkInput,
+    uploadedByAccountId: string,
+  ): Promise<string[]> {
+    const promiseFileUpload = attachmentsInput.attachments
+    const listFileId: string[] = []
+    if (promiseFileUpload) {
+      const arrFileId = promiseFileUpload.map(async (document) => {
+        const { createReadStream, filename, encoding } = await document
+
+        // eslint-disable-next-line no-console
+        console.log('encoding', encoding)
+
+        const documentFile = await this.fileStorageService.uploadFromReadStream(
+          {
+            orgId,
+            originalFileName: filename,
+            readStream: createReadStream(),
+            uploadedByAccountId,
+          },
+        )
+
+        return documentFile.id
+      })
+
+      await Promise.all(arrFileId)
+        .then((fileIds) => {
+          fileIds.forEach((fileId) => {
+            listFileId.push(fileId)
+          })
+        })
+        .catch((err) => {
+          throw new Error(err)
+        })
+    }
+
+    return listFileId
   }
   /**
    * END CLASSWORK ASSIGNMENT
