+++ packages/server/src/modules/classwork/classworkAssignments.resolver.ts
@@ -1,29 +1,14 @@
-import {
-  forwardRef,
-  Inject /** , UsePipes, ValidationPipe */,
-  UsePipes,
-  ValidationPipe,
-} from '@nestjs/common'
-import {
-  Args,
-  ID,
-  Mutation,
-  /* Parent, */
-  Query,
-  /* ResolveField, */
-  Resolver,
-} from '@nestjs/graphql'
+import { UsePipes, ValidationPipe } from '@nestjs/common'
+import { Args, ID, Mutation, Query, Resolver } from '@nestjs/graphql'
 // import { differenceInMinutes } from 'date-fns'
 import { DocumentType } from '@typegoose/typegoose'
 
 // eslint-disable-next-line import/order
 import { CurrentAccount, CurrentOrg, Publication, UseAuthGuard } from 'core'
-import { AuthService } from 'modules/auth/auth.service'
 import { P } from 'modules/auth/models'
 // eslint-disable-next-line import/order
-import { FileStorageService } from 'modules/fileStorage/fileStorage.service'
 import { Org } from 'modules/org/models/Org'
-import { ANY, Nullable, PageOptionsInput } from 'types'
+import { Nullable, PageOptionsInput } from 'types'
 
 import { ClassworkService } from './classwork.service'
 import {
@@ -37,12 +22,7 @@ import { ClassworkAssignment } from './models/ClassworkAssignment'
 
 @Resolver((_of) => ClassworkAssignment)
 export class ClassworkAssignmentsResolver {
-  constructor(
-    private readonly classworkService: ClassworkService,
-    @Inject(forwardRef(() => AuthService))
-    private readonly authService: AuthService,
-    private readonly fileStorageService: FileStorageService,
-  ) {}
+  constructor(private readonly classworkService: ClassworkService) {}
 
   /**
    *START ASSIGNMENTS RESOLVER
@@ -139,49 +119,18 @@ export class ClassworkAssignmentsResolver {
 
   @Mutation((_returns) => ClassworkAssignment)
   @UseAuthGuard(P.Classwork_AddAttachmentsToClassworkAssignment)
-  async addAttachmentsToClassworkAssignments(
+  async addAttachmentsToClassworkAssignment(
     @CurrentOrg() org: Org,
     @CurrentAccount() account: Account,
     @Args('classworkAssignmentId', { type: () => ID })
     classworkAssignmentId: string,
     @Args('attachmentsInput') attachmentsInput: AddAttachmentsToClassworkInput,
   ): Promise<Nullable<DocumentType<ClassworkAssignment>>> {
-    const promiseFileUpload = attachmentsInput.attachments
-    const listFileId: ANY[] = []
-    if (promiseFileUpload) {
-      const arrFileId = promiseFileUpload.map(async (document) => {
-        const { createReadStream, filename, encoding } = await document
-
-        // eslint-disable-next-line no-console
-        console.log('encoding', encoding)
-
-        const documentFile = await this.fileStorageService.uploadFromReadStream(
-          {
-            orgId: org.id,
-            originalFileName: filename,
-            readStream: createReadStream(),
-            uploadedByAccountId: account.id,
-          },
-        )
-
-        return documentFile.id
-      })
-
-      await Promise.all(arrFileId)
-        .then((fileIds) => {
-          fileIds.forEach((fileId) => {
-            listFileId.push(fileId)
-          })
-        })
-        .catch((err) => {
-          throw new Error(err)
-        })
-    }
-    //
     return this.classworkService.addAttachmentsToClassworkAssignment(
       org.id,
       classworkAssignmentId,
-      listFileId,
+      attachmentsInput,
+      account.id,
     )
   }
