+++ packages/server/src/modules/classwork/classworkMaterial.resolver.ts
@@ -1,13 +1,11 @@
-import { forwardRef, Inject, UsePipes, ValidationPipe } from '@nestjs/common'
+import { UsePipes, ValidationPipe } from '@nestjs/common'
 import { Args, ID, Mutation, Query, Resolver } from '@nestjs/graphql'
 import { DocumentType } from '@typegoose/typegoose'
 
 import { CurrentAccount, CurrentOrg, Publication, UseAuthGuard } from 'core'
-import { AuthService } from 'modules/auth/auth.service'
 import { P } from 'modules/auth/models'
-import { FileStorageService } from 'modules/fileStorage/fileStorage.service'
 import { Org } from 'modules/org/models/Org'
-import { ANY, Nullable, PageOptionsInput } from 'types'
+import { Nullable, PageOptionsInput } from 'types'
 
 import { ClassworkService } from './classwork.service'
 import {
@@ -21,12 +19,7 @@ import { ClassworkMaterial } from './models/ClassworkMaterial'
 
 @Resolver((_of) => Classwork)
 export class ClassworkMaterialResolver {
-  constructor(
-    private readonly classworkService: ClassworkService,
-    @Inject(forwardRef(() => AuthService))
-    private readonly authService: AuthService,
-    private readonly fileStorageService: FileStorageService,
-  ) {}
+  constructor(private readonly classworkService: ClassworkService) {}
 
   /**
    *START MATERIAL RESOLVER
@@ -143,42 +136,11 @@ export class ClassworkMaterialResolver {
     classworkAssignmentId: string,
     @Args('attachmentsInput') attachmentsInput: AddAttachmentsToClassworkInput,
   ): Promise<Nullable<DocumentType<ClassworkMaterial>>> {
-    const promiseFileUpload = attachmentsInput.attachments
-    const listFileId: ANY[] = []
-    if (promiseFileUpload) {
-      const arrFileId = promiseFileUpload.map(async (document) => {
-        const { createReadStream, filename, encoding } = await document
-
-        // eslint-disable-next-line no-console
-        console.log('encoding', encoding)
-
-        const documentFile = await this.fileStorageService.uploadFromReadStream(
-          {
-            orgId: org.id,
-            originalFileName: filename,
-            readStream: createReadStream(),
-            uploadedByAccountId: account.id,
-          },
-        )
-
-        return documentFile.id
-      })
-
-      await Promise.all(arrFileId)
-        .then((fileIds) => {
-          fileIds.forEach((fileId) => {
-            listFileId.push(fileId)
-          })
-        })
-        .catch((err) => {
-          throw new Error(err)
-        })
-    }
-    //
     return this.classworkService.addAttachmentsToClassworkMaterial(
       org.id,
       classworkAssignmentId,
-      listFileId,
+      attachmentsInput,
+      account.id,
     )
   }
